app/requirements.txt
# =============================================================================
# Requirements pour l'architecture MVVM - Outil Emailing
# =============================================================================

# Framework Web
flask==3.0.0
flask-cors==4.0.0

# Serveur de production
gunicorn==21.2.0

# Configuration
python-dotenv==1.0.0

# Utilitaires HTTP
requests==2.31.0

# Utilitaires (optionnel pour le développement)
# watchdog==3.0.0  # Pour le hot-reload en dev
